Basic proposal cancellation
Model field "cancelled" added. The proposal update pages get an
extra form that posts to a proposal cancel view, updating the
"cancelled" flag.

If a proposal is cancelled, its update page has only one button
to re-activate the proposal. After re-activation, the proposal
can then be updated again.

TODO: Position of the cancel/re-activate button needs to be taken
care of when the real UI is implemented.