Fix accidentally miss proposals as daily cron job
Now the behavior has changed, so has the command options. Summary of
all changes:

- All operations are timezone aware, and hard-coded to use Taiwan
  timezone.
- Add an hour option which should match the hour when the cron job runs.

The detail about the behavior change is that daily cron jobs guarantee
to run after the specified time, say at 21:00 everyday, but it won't be
the exact same time. Previously, we compute the exact 24 x N hours to
find the proposal, so we might lose some talks submitted around 21:00.

For example, if the cron job is set to run at 21:00, with command
options being `--days 3 --hour 21`. An example of the running schedule
can be

    Nth Job | Actual run time | Report proposals submitted during
    1st     | Jan 10 21:15    | Jan  8 21:00 - Jan 10 20:59
    2nd     | Jan 13 21:34    | Jan 10 21:00 - Jan 13 21:00

If the same command is called on 20:30, it failed with a CommandError
because 21:00 has yet to come.

Surprise hidden in the comment: a set of unittest based on test time has
been created. Welcome to the stochastic and unpredictable failing tests
:)