Merge review create and update views
Since review has one-to-one (optional) relation for each user at a
certain time, we can simply use the proposal pk as the key in the
URL scheme, and infer other values automatically. This simplifies
both the scheme and logic inside views.

The above relation is now promised at db level through a
unique_together constraint. This actually should have been enforced
even with the previous implementation but wasn't. Anyway it is now.