Toggle proposal creation/edit with settings
Two additional settings:

PROPOSALS_CREATABLE: If true, creation of proposals are allowed.
PROPOSALS_EDITABLE: If true, edit/cancellation of proposals are allowed.

If either of the settings is false, related links will be hidden, and
views return 404.

This commit sets both settings as false, as needed for the review process.