Un-escape HTML in JavaScript, don't use mark_safe
This fixes an XSS problem when displaying rendered Markdown fields.
If the user puts a raw <script> tag inside the field, it will be
rendered as raw HTML, thus allowing XSS attacks.

The current implementation removes mark_safe on the server side,
and uses DOMParser to safely parse the escaped text into raw HTML
(without executing it), and then feed it to marked.js. The output
is safe since marked.js performs HTML sanitation automatically.

Thanks for @ccwang002 for reporting this issue.