Add warning when leaving dirty proposal form
Use jQuery-AreYouSure to make things easy. Most modern browsers do
not allow dialog message customisation now, so I didn't bother to translate
the prompt.

Vendored SimpleMDE is upgraded to 1.11.2. This incorporates the
`forceSync` option introduced in 1.10.1. We need this to make the
texture dirty detection work, so we can display the onbeforeunload
warning dialog correctly. `forceSync` is set to true during SimpleMDE
setup so jQuery-AreYouSure can detect changes.

Note: The $.areYouSure call needs to be after SimpleMDE is configured,
so we are listening to the right elements for events.

Close #317.