Add instructions for running production server
These instructions explain how to do a container-based deployment of
production server.
Container-based deployment is chosen because it is easier to version
control the application environment and greatly simplifies the building
step.
While this makes actually running the server more verbose (since the
environment variables are given directly on command-line), there is
merit in being explicit in the production environment.