Add Dockerfile
This Dockerfile builds a container image with virtualenv enabled, with
additional utility (wait-for-it) preinstalled.
The main user running the application is called 'docker', without an
associated group.
Python files are pre-compiled with '-m compileall' module to make
startup faster.
The file 'dbfnew.py' is ignored in compileall with `-x 'dbfpy/dbfnew\.py'`
because it contains Python 2 syntax.