Fix navbar display issue in safari
I have no idea why I made the navbar as flex at the first place.
Maybe I was just trying to make "My PyCon" rightmost, and try to
do that by making the nav items and "My PyCon" as flex items.

But in the end, "My PyCon" remains as an item inside the nav,
which makes it only one element inside the flex, and this cause
a problem in Safari, making the entire flex container with no
height, so the display is wrong and you cannot hit any item on the
menu.