Make proposals and reviews state editable in admin
The configuration of proposals and reviews were written in the
Django's settings code, and changing them requires a change
of the source code, which is not convenient.

In this commit, the configurations are stored in the database,
and the concept is similar to windows registry. A third party
library dj-registry is used here.

The following are the corresponding variables:

PROPOSALS_CREATABLE -> pycontw-2020.proposals.creatable
PROPOSALS_EDITABLE -> pycontw-2020.proposals.editable
PROPOSALS_WITHDRAWABLE -> pycontw-2020.proposals.withdrawable
REVIEWS_STAGE -> pycontw-2020.reviews.stage
REVIEWS_VISIBLE_TO_SUBMITTERS ->
pycontw-2020.reviews.visible.to.submitters

And there is also one addition key which is
pycontw-2020.proposals.disable.after

You can specify a datetime formatted as 2020-04-30 23:59:59-1200.
After settings this, proposal creatable, editable and withdrawable
would all be false if current time passes the time set. This
has higher priority.