Fix wrong default review form values
When we are in review stage 2, it will try to retrieve the previous
review data and fill in to the fields as default values.

First, it tries to fetch if there are reviews in the same stage. If not,
it will try to fetch if there are reviews in the previous stage.

The first one does not work because it looked at the wrong parameter.
It should be looking for 'instance' instead of 'object' in the form
kwargs of the Django generic ModelFormMixin. Fixing this fixed the issue
and it will display the correct default value.