Use BigAutoField for User.id
We have been using the bigint data type for User.id since ab2573d6a25a, but the
id field that Django automatically create for us uses AutoField, which maps to
the int data type.
This mismatch in data type is mostly benign, but when User.id grow large
enough, it will blowup when we try to use setval on User.id (e.g. importing
fixture with `manage.py loaddata`), with an error complaining value out of
range.