fix(user): return github & twitter profile url only when id is given
[WHY]
For those users that did not provide github/twitter id, the response of event APIs gives only the hostname of social media (e.g. https://github.com). An additional function for determining whether the user provides the profile id is unwanted but inevitably required.

[HOW]
Returning empty string if the user do not provide the profile id.